fix: fixed some bugs with adding html template respond logics

+++ api/controller/subscribe.py
@@ -1,12 +1,15 @@
-"""
-Subscribe endpoints handler
-"""
-
 import os
 from typing import Any, Dict, Optional, List
 
 from repository.subscription_repository import repo
 from utils.error import SubscriptionNotFound
+from utils.token import generate_unsubscribe_token
+from utils.response import html_response
+from templates.template import get_unsubscribe_error_page, get_unsubscribe_success_page
+
+import logging
+logger = logging.getLogger("[SSU_ANNOUNCEMENT_API_CONTROLLER]")
+logger.setLevel(logging.INFO)
 
 CATEGORIES: List[str] = [
     "ssu_path",
@@ -33,7 +36,7 @@ def get_subscribes(event: Dict[str, Any], db_session: Any) -> Dict[str, Any]:
 
     return {"subscribes": subscribes}
 
-def add_subscribe(event: Dict[str, Any], db_session: Any) -> None:
+def subscribe(event: Dict[str, Any], db_session: Any) -> None:
     # Query parameter에서 Email, Category 추출
     email: str = _get_email_from_event(event)
     category: str = _get_category_from_event(event)
@@ -41,19 +44,42 @@ def add_subscribe(event: Dict[str, Any], db_session: Any) -> None:
     # Set table
     table: Any = db_session.Table(os.getenv("SUBSCRIPTION_TABLE"))
 
-    # Add subscription
-    repo.add_subscription(table=table, email=email, category=category)
+    # Generate subscription unique token for unsubscribe
+    token: str = generate_unsubscribe_token(email=email, category=category)
 
-def delete_subscribe(event: Dict[str, Any], db_session: Any) -> None:
-    # Query parameter에서 Email, Category 추출
-    email: str = _get_email_from_event(event)
-    category: str = _get_category_from_event(event)
+    # Add subscription with unique token
+    repo.add_subscription(table=table, email=email, category=category, token=token)
 
-    # Set table
-    table: Any = db_session.Table(os.getenv("SUBSCRIPTION_TABLE"))
+def unsubscribe(event: Dict[str, Any], db_session: Any) -> Dict[str, Any]:
+    try:
+        # Get query parameters
+        token: str = _get_token_from_event(event)
 
-    # Delete subscription
-    repo.delete_subscription(table=table, email=email, category=category)
+        # Validate parameters
+        if not token:
+            return html_response(get_unsubscribe_error_page("잘못된 요청입니다. 이메일에서 제공된 링크를 사용해주세요."), 400)
+
+        # Set table
+        table: Any = db_session.Table(os.getenv("SUBSCRIPTION_TABLE"))
+
+        # Verify token
+        # bool, dict[str, Any]
+        success, data = repo.verify_unsubscribe_token(table, token)
+        logger.info(f"Verify token: {success}, {data}")
+        if not success:
+            return html_response(get_unsubscribe_error_page("유효하지 않은 요청입니다. 토큰이 만료되었거나 올바르지 않습니다."), 400)
+
+        # Delete subscription
+        email: str = data.get("Email")
+        category: str = data.get("Category")
+        table: Any = db_session.Table(os.getenv("SUBSCRIPTION_TABLE"))
+        repo.delete_subscription(table=table, email=email, category=category)
+
+        # Return success HTML
+        return html_response(get_unsubscribe_success_page(email, category), 200)
+
+    except Exception as e:
+        return html_response(get_unsubscribe_error_page(f"오류가 발생했습니다"), 500)
 
 def _get_email_from_event(event: Dict[str, Any]) -> str:
     email: Optional[str] = event.get("queryStringParameters", {}).get("email", "")
@@ -70,3 +96,9 @@ def _get_category_from_event(event: Dict[str, Any]) -> str:
         raise ValueError("Invalid category")
 
     return category
+
+def _get_token_from_event(event: Dict[str, Any]) -> str:
+    token: Optional[str] = event.get("queryStringParameters", {}).get("token", "")
+    if not token:
+        raise ValueError("Token parameter is required")
+    return token
